Add sealing of contexts
Now there is a clearer separation between the "building" phase, when you
add new units to a context, and the "use" phase, when you start to do
things with that context that you built up.  To transition from the
building phase to the use phase, you must invoke the `seal` method.
Once you've done that, you cannot add any addition units to the context.